chore: add resolveAuthConfig
